requirements.txt
# ==========================================
# Core Dependencies
# ==========================================

# Web Server
fastapi
uvicorn[standard]
python-multipart
websockets

# PDF Processing
pypdf2

# Office Document Processing (for conversion to PDF)
python-pptx         # PowerPoint files
python-docx         # Word documents
openpyxl            # Excel spreadsheets
reportlab           # PDF generation from text

# LLM APIs
google-genai  # Gemini 2.5 Pro SDK

# Cloud Storage
google-cloud-storage==2.14.0
google-auth==2.27.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Utilities
python-dotenv
pydantic
